Added a TileManager Script and made the tile move towards the camera
I created a TileManager Script and I made the Tile move towards the camera.
The way I did it, is by creating an empty game object as the child of the camera, then used Vector3 MoveTowards and Lerp to make the Tile move smoothly towards the empty object at a chosen speed.