Basic Tile Generation
Moved the Tile Manager script to the TileManager empty object which I have set as a Child of the Tile Object also used the TileManager Empty object to check for collision when player is near the edge of the tile, so it spawns a new Tile in front. Also, I created a new object behind the player with its own script that destroys the tiles.